Don't attempt to clean up :refer :all requires
Without loading the project, refactor-nrepl's prune-dependencies always
removes :refer :all requires as it doesn't understand which symbols are
present in the 'required' ns.

This change avoids ever removing :refer :all requires, so instead of
erroneously removing these we may sometimes erroneously let them remain.
The latter is preferably as it doesn't break your project :)

This change can be reversed if we find the right way to load the project
ns's before cleaning.

Closes #1